break out traits to fix Action vs ActionRequest usage
